Revert "divide module"
